+++ CHANGELOG.md
@@ -12,6 +12,7 @@ This new version contains significant breaking changes to the graph database and
 - **BREAKING** Moved some CLI commands around to try to create a more consistent interface. `graph canned` is gone and all commands now exist under `find`. Some `find` subcommands moved to `list`.
 - **BREAKING** Changed the way prerequisistes are stored in the meta database
 - **BREAKING** Made `FileStore` require `Send` and `Sync`
+- **BREAKING** `meta set-progress` now sets to true and `clear-progress` added to set to false
 - Create Python bindings for the databases and the application server
 - Exposed some functionality that was not previous `pub` or was just `pub (crate)`
 - Removed some restructions on `find-callers`